+++ packages/transaction-pay-controller/src/types.ts
@@ -1,9 +1,14 @@
-import type { AssetsControllerGetStateForTransactionPayAction } from '@metamask/assets-controller';
+import type {
+  AssetsControllerGetStateForTransactionPayAction,
+  AssetsControllerStateChangeEvent,
+} from '@metamask/assets-controller';
 import type {
   CurrencyRateControllerGetStateAction,
+  CurrencyRateStateChange,
   TokenBalancesControllerGetStateAction,
 } from '@metamask/assets-controllers';
 import type { TokenRatesControllerGetStateAction } from '@metamask/assets-controllers';
+import type { TokenRatesControllerStateChangeEvent } from '@metamask/assets-controllers';
 import type { TokensControllerGetStateAction } from '@metamask/assets-controllers';
 import type { AccountTrackerControllerGetStateAction } from '@metamask/assets-controllers';
 import type { ControllerStateChangeEvent } from '@metamask/base-controller';
@@ -82,7 +87,10 @@ export type AllowedActions =
   | TransactionControllerUpdateTransactionAction;
 
 export type AllowedEvents =
+  | AssetsControllerStateChangeEvent
   | BridgeStatusControllerStateChangeEvent
+  | CurrencyRateStateChange
+  | TokenRatesControllerStateChangeEvent
   | TransactionControllerStateChangeEvent
   | TransactionControllerUnapprovedTransactionAddedEvent;